Moved to IB API 10.19.01
- also added Interl RDFP Mathlib v.20u2 (linux only)

+++ IbTradeQt/Shared/bar.h
@@ -1,4 +1,4 @@
-/* Copyright (C) 2018 Interactive Brokers LLC. All rights reserved. This code is subject to the terms
+/* Copyright (C) 2019 Interactive Brokers LLC. All rights reserved. This code is subject to the terms
  * and conditions of the IB API Non-Commercial License or the IB API Commercial License, as applicable. */
 
 #pragma once
@@ -6,6 +6,7 @@
 #define TWS_API_CLIENT_BAR_H
 
 #include <string>
+#include "Decimal.h"
 
 struct Bar
 {
@@ -14,8 +15,8 @@ struct Bar
 	double low;
 	double open;
 	double close;
-	double wap;
-	long long volume;
+	Decimal wap;
+	Decimal volume;
 	int count;
 };
